+++ examples/hcl/ubuntu/etc/scripts/060-virtualbox.sh
@@ -1,0 +1,1 @@
+../../../../_common/virtualbox.sh