+++ examples/hcl/ubuntu/etc/scripts/080-parallels.sh
@@ -1,0 +1,1 @@
+../../../../_common/parallels.sh