+++ hcl2template/types.packer_config.go
@@ -203,7 +203,7 @@ func (cfg *PackerConfig) getCoreBuildProvisioners(source SourceBlock, blocks []*
 	var diags hcl.Diagnostics
 	res := []packer.CoreBuildProvisioner{}
 	for _, pb := range blocks {
-		if pb.OnlyExcept.Skip(source.Type + "." + source.Name) {
+		if pb.OnlyExcept.Skip(source.String()) {
 			continue
 		}
 		provisioner, moreDiags := cfg.startProvisioner(source, pb, ectx)
@@ -250,7 +250,7 @@ func (cfg *PackerConfig) getCoreBuildPostProcessors(source SourceBlock, blocks [
 	var diags hcl.Diagnostics
 	res := []packer.CoreBuildPostProcessor{}
 	for _, ppb := range blocks {
-		if ppb.OnlyExcept.Skip(source.Type + "." + source.Name) {
+		if ppb.OnlyExcept.Skip(source.String()) {
 			continue
 		}
 
@@ -308,10 +308,11 @@ func (cfg *PackerConfig) GetBuilds(opts packer.GetBuildsOptions) ([]packer.Build
 				continue
 			}
 			src.addition = from.addition
+			src.LocalName = from.LocalName
 
 			pcb := &packer.CoreBuild{
 				BuildName: build.Name,
-				Type:      src.Ref().String(),
+				Type:      src.String(),
 			}
 
 			// Apply the -only and -except command-line options to exclude matching builds.